requirements.txt
pygame
PyYAML